Remove `utils.rs` module
